pkp/pkp-lib#9744 Rename TableNext component to Table
