+++ src/pages/dashboard/components/DashboardTable/CellReviewAssignmentActions.vue
@@ -13,7 +13,7 @@
 <script setup>
 import {defineProps} from 'vue';
 import PkpButton from '@/components/Button/Button.vue';
-import TableCell from '@/components/TableNext/TableCell.vue';
+import TableCell from '@/components/Table/TableCell.vue';
 import {useDashboardPageStore} from '@/pages/dashboard/dashboardPageStore.js';
 import {useLocalize} from '@/composables/useLocalize';
