+++ src/pages/dashboard/components/DashboardTable/CellReviewAssignmentActivity.vue
@@ -20,7 +20,7 @@
 
 <script setup>
 import {defineProps, computed} from 'vue';
-import TableCell from '@/components/TableNext/TableCell.vue';
+import TableCell from '@/components/Table/TableCell.vue';
 import CellReviewAssignmentActivityAlert from './CellReviewAssignmentActivity/CellReviewAssignmentActivityAlert.vue';
 import {useDashboardPageStore} from '@/pages/dashboard/dashboardPageStore';
