+++ src/pages/dashboard/components/DashboardTable/CellSubmissionDays.vue
@@ -3,7 +3,7 @@
 </template>
 <script setup>
 import {defineProps} from 'vue';
-import TableCell from '@/components/TableNext/TableCell.vue';
+import TableCell from '@/components/Table/TableCell.vue';
 
 defineProps({
 	item: {type: Object, required: true},
