+++ src/pages/userInvitation/UserInvitationUserGroupsTable.vue
@@ -100,12 +100,12 @@
 <script setup>
 import {computed} from 'vue';
 import {useTranslation} from '@/composables/useTranslation';
-import PkpTable from '@/components/TableNext/Table.vue';
-import TableCell from '@/components/TableNext/TableCell.vue';
-import TableHeader from '@/components/TableNext/TableHeader.vue';
-import TableColumn from '@/components/TableNext/TableColumn.vue';
-import TableBody from '@/components/TableNext/TableBody.vue';
-import TableRow from '@/components/TableNext/TableRow.vue';
+import PkpTable from '@/components/Table/Table.vue';
+import TableCell from '@/components/Table/TableCell.vue';
+import TableHeader from '@/components/Table/TableHeader.vue';
+import TableColumn from '@/components/Table/TableColumn.vue';
+import TableBody from '@/components/Table/TableBody.vue';
+import TableRow from '@/components/Table/TableRow.vue';
 import FieldSelect from '@/components/Form/fields/FieldSelect.vue';
 import PkpButton from '@/components/Button/Button.vue';
 import FieldText from '@/components/Form/fields/FieldText.vue';
